Memory system and Buddy system
still have bug when use async io and memory system simultaneously.

+++ Lab4/kernel/include/string.h
@@ -6,11 +6,11 @@ char* char_to_hex_str(char c);
 char* int_to_hex_str(int i);
 char* long_to_hex_str(long l);
 
-uint32_t hex_str_to_uint(char *s);
+uint64_t hex_str_to_uint(char *s);
 
 char* uint_to_dec_str(uint64_t i);
 
-uint32_t dec_str_to_uint(char *s);
+uint64_t dec_str_to_uint(char *s);
 
 int strcmp(const char *s1, const char *s2);
 int strncmp(const char *s1, const char *s2, uint32_t len);
